+++ src/lib/harness/register.ts
@@ -2,6 +2,7 @@ import { ensureClaudeRegistered } from "./claudeAdapter";
 import { ensureCodexRegistered } from "./codexAdapter";
 import { ensureCursorRegistered } from "./cursorAdapter";
 import { ensureFxRegistered } from "./fxAdapter";
+import { ensureGrokRegistered } from "./grokAdapter";
 import { ensureOpenCodeRegistered } from "./opencodeAdapter";
 import { ensureOmpRegistered } from "./ompAdapter";
 import { ensurePiRegistered } from "./piAdapter";
@@ -11,6 +12,7 @@ export function registerBuiltinHarnesses(): void {
   ensureClaudeRegistered();
   ensureCursorRegistered();
   ensureCodexRegistered();
+  ensureGrokRegistered();
   ensureOpenCodeRegistered();
   ensurePiRegistered();
   ensureOmpRegistered();